+++ src/Sql/ConnectionExtensions.cs
@@ -12,11 +12,11 @@ public static partial class ConnectionExtensions {
 	/// Creates a new command associated with the specified connection.
 	/// </summary>
 	/// <param name="connection">The connection to the data source.</param>
-	/// <param name="sql">The SQL query to be executed.</param>
+	/// <param name="text">The SQL query to be executed.</param>
 	/// <param name="parameters">The parameters of the SQL query.</param>
 	/// <param name="options">The command options.</param>
 	/// <returns>The newly created command.</returns>
-	public static IDbCommand CreateCommand(this IDbConnection connection, string sql, ParameterCollection? parameters = null, CommandOptions? options = null) {
+	public static IDbCommand CreateCommand(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null) {
 		var command = connection.CreateCommand();
 		command.CommandText = sql;
 		command.CommandTimeout = options?.Timeout ?? 30;